TinyEKFGen.jl

A Julia C code generator for use with TinyEKF.

Index

TinyEKFGen.diffvwrtv

TinyEKFGen.outputHeader

TinyEKFGen.diffvwrtv — Method

function diffvwrtv(e::Array{Basic,1}, v::Array{Basic,1})

Compute Jacobian of a vector of expressions e with respect to another vector of expressions v.

TinyEKFGen.outputHeader — Function

function outputHeader(filename, state, px, hx, constants=Dict(),post=Dict())
function outputHeader(filename, state, px, F, hx, H,constants=Dict(),post=Dict())

filename: output C header file name
state: Kalman filter state
px:    Predicted new state f(state=x)
F:     df/dx, state update matrix
hx:    Predicted measurements from current state, h(state=x)
H:     dh/dx, observation derivative matrix
constants: Map from symbol name to value for symbols to be treated as constants
post:  map of matrix name to matrix to be printed at the end of the output code